Resolution.Target => ResolutionTarget, and @Nullable change
